Update dependency matrix and tests
Remove tests for Puppet 4.3.0 and earlier. Test failures appear to
indicate that module functionality uses features that are not available
in puppet 4.3.0 and earlier.

Increased upper limit on the puppetlabs/concat support to < 5.0.0. As
the tests are already testing against concat master, we know it's
compatible with concat 4.x.

Added Puppet 5.x support to the metadata.json file. Been testing using
Puppet 5.0.1 since I started on this PR.