Refactor components and config to be more dynamic
Removed the concat from the config class and created a new
defined type called generate_concat. This defined type will take in
2 parameters - confdir and target - and create a concat resource off
those values if the resource does not already exist.

The component defines have been refactored to use this new defined type
and the config classes will set the default for confdir to $::rsyslog::confdir
if no other directory is specified.