Make rules optional and ensure that stop or rules has content
Upon speaking with some Rsyslog engineers, I found out that a
ruleset can be empty except for a stop within it. To allow for
this, I made rules and stop optional parameters, then added a check
that will raise a "fail" if rules is empty AND stop is false.

Associated test added as well.