Restart rsyslog if feature_packages are installed
Setup rsyslog so that if any $feature_packages are given rsyslog will
restart after those $feature_packages are installed.

Forcing a restart of rsyslog after installation of any $feature_packages
are installed is important if you have an rsyslog config that is not
valid without the additional package from $feature_packages in place.